api/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
httpx==0.28.1
pydantic==2.10.4
google-genai==1.16.1
python-dotenv==1.0.1
